🎨

+++ pages/api/linebot/usecases/delete-task.ts
@@ -5,6 +5,15 @@ import { prisma } from '~/libs/prisma'
 export const deleteTaskUsecase = async (event: PostbackEvent) => {
   const [_, __, taskId] = event.postback.data.split('_')
 
+  const task = await prisma.task.findUnique({ where: { id: parseInt(taskId, 10) } })
+  if (task === null) {
+    await lineClient.replyMessage(event.replyToken, {
+      type: 'text',
+      text: `すでに削除されています`
+    })
+    return
+  }
+
   await prisma.task.delete({
     where: {
       id: parseInt(taskId, 10)
@@ -13,6 +22,6 @@ export const deleteTaskUsecase = async (event: PostbackEvent) => {
 
   await lineClient.replyMessage(event.replyToken, {
     type: 'text',
-    text: `${taskId}を削除しました`
+    text: `「${task.title}」を削除しました`
   })
 }
